refactor question components
