+++ apps/frontend/src/components/questions/questionPage/DescriptionPanel.tsx
