+++ apps/frontend/src/components/questions/questionTable/EditQuestionModal.tsx
@@ -2,8 +2,8 @@ import ReactMarkdown from "react-markdown";
 import { Complexity } from "@/utils/enums/Complexity";
 import { Category } from "@/utils/enums/Category";
 import { useEffect, useState } from "react";
-import Modal from "../Modal";
 import { Question } from "@/database/question/entities/question.entity";
+import Modal from "@/components/Modal";
 
 type EditQuestionModalProps = {
   onEditQuestion: {
