+++ apps/frontend/src/pages/questions/index.tsx
@@ -1,10 +1,10 @@
 import { useEffect, useState } from "react";
 import { PlusIcon } from "@heroicons/react/20/solid";
-import QuestionTable from "@/components/questions/QuestionTable";
 import useQuestion from "@/hook/useQuestions";
 import LoadingModal from "@/components/LoadingModal";
 import useSessionUser from "@/hook/useSessionUser";
 import { Role } from "@/utils/enums/Role";
+import QuestionTable from "@/components/questions/questionTable/QuestionTable";
 
 export default function QuestionsRepo() {
   const [openAdd, setOpenAdd] = useState(false);
